+++ addOns/automation/CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Unreleased
 ### Added
 - Add Save As button that allows user to save the automation plan to a different file (Issue 7178).
+- Support for monitor tests.
 
 ### Changed
 - Maintenance changes.